+++ src/thordata/__init__.py
@@ -5,7 +5,7 @@ Official Python client for Thordata's Proxy Network, SERP API,
 Universal Scraping API (Web Unlocker), and Web Scraper API.
 """
 
-__version__ = "1.6.0"
+__version__ = "1.8.0"
 __author__ = "Thordata Developer Team/Kael Odin"
 __email__ = "support@thordata.com"